examples/requirements.txt
kafka-python==2.0.2
confluent-kafka==2.3.0
aiokafka==0.10.0 
aws-msk-iam-sasl-signer-python==1.0.2